keyboard shortcuts: add jk, Enter, Esc for news item navigation
in addition to the arrow keys, this allows to use the jk keys to be
used to navigate through the news items.
These keys are widely used for navigation and allow the fingers to rest
at their natural position on the keyboard.
Using the Escape to exit the news item detail view is intuitive
and using enter to open the item as well.

gui shortcuts for search function

/ to focus the search bar
When in the search bar:
Esc clears the text
With another Esc if the field is empty un-sets the focus, allows the
other shortcuts to work

previous:
pressing arrow up/down or j/k to navigate, with a news item detail open,
navigates through the new item list which located invisible in the back
now:
arrow up/dow or j/k scrolls the news item text which is visible

Use hjkl for up/down/open/close
use u/U for upvote/downvote (like/unlike)
use g/G for group/ungroup

gui shortcuts: update the shortcuts docs