Feature/bots collectors cleanup (#23)
Rewrite Collector and Bot, Parameter & Parameter Value.

With this new approach, parameters are shared, and not every parameter value has its own parameter.

For example:

The "REFRESH_INTERVAL" each Bot and Collector has, the exact same row (every column identical except the id) existed up to 20 times, now it gets pre seeded and exists only once.

To achieve this, I switched the primary key from the artificial "id" to the natural "key".

This means Bots and Collectors do not need to publish their own parameters anymore, rather parameters should be created separately and are only accessed. This is a necessary step to switch to a publish subscribe worker setup.