Feature/rewrite acl (#150)
This PR Fixes #47 #86

ACLEntry has been renamed to "RoleBasedAccess", with the goal to in the feature make it easier to implement "real" RBAC as a access control in Taranis AI, the changes in this PR should lay the foundation in this direction.

Features:
An Administrator with the correct permssion:

  CONFIG_ACL_ACCESS
  CONFIG_ACL_CREATE
  CONFIG_ACL_UPDATE
  CONFIG_ACL_DELETE
Can update or create ACLs under /config/acls.

He has to provide an item type, item id, unique name, and a role to assign the acl to.

For backwards compatibility ACLs are only active if there is at least on ACL for a specific item_type.
This information is shown on top of the ACL page.

If there is an ACL for a specific type only these roles are allowed to read, and if 'write' flag is set, also write the given resource.