Remove language validation with langcodes from `frontend` (#729)
Because now we are doing parsing in the
[worker](https://github.com/taranis-ai/taranis-ai/blob/master/src/worker/worker/types.py#L72-L87),
it is not necessary to validate the language field in the frontend.

One edge case would come up, and that is with the manual news items.
This might be especially important if we allow the JSON file as input -
then it should go through the same sanitisation as in the worker (not
only the Language field)

## Summary by Sourcery

Remove frontend language normalization for news items now that language
parsing is handled in the worker.

Enhancements:
- Drop the NewsItem language field validator that normalized language
codes.
- Remove the unused langcodes dependency from the models package
configuration.